Compiling with warnings with new routes
I changed "api" to the less genetic route "encounter" and added a new
monster_api file that handles a "monster" route that will be used to
retrieve single monster requests. I also renamed files to be more
consistent with the encounter naming. The code is currently compiling
and I am able to get fetch data from the server on the client side.